chore: change packages structures
