+++ feature-characters/build.gradle
@@ -44,5 +44,4 @@ dependencies {
 
     implementation project(':domain')
     implementation project(':platform:core')
-
 }
