+++ feature-characters/src/main/java/paixao/leonardo/marvel/heroes/feature/character/di/CharacterModule.kt
@@ -5,8 +5,8 @@ import org.kodein.di.bind
 import org.kodein.di.instance
 import org.kodein.di.provider
 import paixao.leonardo.marvel.heroes.domain.services.CharactersHandler
-import paixao.leonardo.marvel.heroes.feature.character.CharacterViewModel
-import paixao.leonardo.marvel.heroes.feature.character.CharactersAgent
+import paixao.leonardo.marvel.heroes.feature.character.screens.listing.CharacterViewModel
+import paixao.leonardo.marvel.heroes.feature.character.screens.listing.CharactersAgent
 
 object CharacterModule {
     private const val MODULE_NAME = "feature-character-module"
