+++ feature-characters/src/main/java/paixao/leonardo/marvel/heroes/feature/character/screens/listing/CharacterViewModel.kt
@@ -1,4 +1,4 @@
-package paixao.leonardo.marvel.heroes.feature.character
+package paixao.leonardo.marvel.heroes.feature.character.screens.listing
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.channels.Channel
