+++ feature-characters/src/main/java/paixao/leonardo/marvel/heroes/feature/character/screens/listing/entries/FavoriteCharactersEntry.kt
@@ -1,9 +1,9 @@
-package paixao.leonardo.marvel.heroes.feature.character.entries
+package paixao.leonardo.marvel.heroes.feature.character.screens.listing.entries
 
 import android.view.View
 import com.xwray.groupie.viewbinding.BindableItem
 import paixao.leonardo.marvel.heroes.feature.R
-import paixao.leonardo.marvel.heroes.feature.character.customviews.OnRetry
+import paixao.leonardo.marvel.heroes.feature.character.screens.listing.customviews.OnRetry
 import paixao.leonardo.marvel.heroes.feature.core.exceptions.MarvelException
 import paixao.leonardo.marvel.heroes.feature.databinding.ItemFavoriteCharacterListBinding
