+++ feature-characters/src/main/res/layout/item_character_list.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<paixao.leonardo.marvel.heroes.feature.character.customviews.CharacterView xmlns:android="http://schemas.android.com/apk/res/android"
+<paixao.leonardo.marvel.heroes.feature.character.screens.listing.customviews.CharacterView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -13,4 +13,4 @@
         android:layout_below="@id/tabLayout"
         app:layoutManager="androidx.recyclerview.widget.GridLayoutManager" />
 
-</paixao.leonardo.marvel.heroes.feature.character.customviews.CharacterView>
+</paixao.leonardo.marvel.heroes.feature.character.screens.listing.customviews.CharacterView>
